% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CensoredMediatorCodes.R
\name{CensoredMediator}
\alias{CensoredMediator}
\title{CensoredMediator}
\usage{
CensoredMediator(data, outcome.type, casecontrol, prev, printout)
}
\arguments{
\item{data}{Input dataset file in the .csv format.}

\item{outcome.type}{The type of the outcome in the mediation model. outcome.type="continuous" or "binary".}

\item{casecontrol}{If the study is a case-control study.
If outcome.type="continuous": casecontrol=FALSE
If outcome.type="binary": casecontrol=FALSE if it is not a case-control study, casecontrol=TRUE if it is a case-control study}

\item{prev}{Prevalance of the case in the general population.
Set it as 0 when it is not a case-control study.
Set it as a positive number when it is a case-control study.
If outcome.type="continuous": casecontrol=FALSE; prev=0
If outcome.type="binary": casecontrol=FALSE; prev=0 if it is not a case-control study, casecontrol=TRUE; prev!=0 if it is a case-control study}

\item{printout}{Print the estimated coefficients for different paths.}
}
\value{
list(alpha.hat,beta.opt,IE.results)
}
\description{
Analyze a mediation model with a censored mediator
}
