% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CensoredMediatorCodes.R
\name{loglikelihood}
\alias{loglikelihood}
\title{Title loglikelihood
Description:        Calculate the coefficients for Y~X+M. Called by "myBeta".}
\usage{
loglikelihood(x, data, alpha, tmp, outcome.type, casecontrol)
}
\arguments{
\item{x}{Initial values for parameters to be optimized over.}

\item{data}{Input dataset.}

\item{alpha}{Coefficients estimated for M~X.}

\item{tmp}{AFT model residuals by the Kaplan-Meier estimator.}

\item{outcome.type}{The type of the outcome in the mediation model.  outcome.type="continuous" or "binary".}

\item{casecontrol}{If the study is a case-control study.
If outcome.type="continuous": casecontrol=FALSE
If outcome.type="binary":
          casecontrol=FALSE if it is not a case-control study
          casecontrol=TRUE if it is a case-control study}
}
\value{
negative loglikelihood
}
\description{
Title loglikelihood
Description:        Calculate the coefficients for Y~X+M. Called by "myBeta".
}
