% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CensoredMediatorCodes.R
\name{myBeta}
\alias{myBeta}
\title{Title myBeta
Description:        Calculate the coefficients for Y~X+M. Called by "CensoredMediator".}
\usage{
myBeta(alpha, data, outcome.type, casecontrol)
}
\arguments{
\item{alpha}{Coefficients estimated for M~X.}

\item{data}{Input dataset.}

\item{outcome.type}{The type of the outcome in the mediation model. outcome.type="continuous" or "binary".}

\item{casecontrol}{If the study is a case-control study.
If outcome.type="continuous": casecontrol=FALSE
If outcome.type="binary":  casecontrol=FALSE if it is not a case-control study  casecontrol=TRUE if it is a case-control study}
}
\value{
list(beta.opt,tmp)
}
\description{
Title myBeta
Description:        Calculate the coefficients for Y~X+M. Called by "CensoredMediator".
}
