% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CensoredMediatorCodes.R
\name{mybootstrap}
\alias{mybootstrap}
\title{mybootstrap}
\usage{
mybootstrap(
  data_hold,
  IE.star,
  boot,
  outcome.type,
  casecontrol,
  prev,
  printout
)
}
\arguments{
\item{data_hold}{Input dataset.}

\item{IE.star}{Estimated IE and PM, obtained from "CensoredMediator".}

\item{boot}{Number of bootstrap samples.}

\item{outcome.type}{The type of the outcome in the mediation model. outcome.type="continuous" or "binary".}

\item{casecontrol}{If the study is a case-control study.
If outcome.type="continuous": casecontrol=FALSE
If outcome.type="binary":
         casecontrol=FALSE if it is not a case-control study
          casecontrol=TRUE if it is a case-control study}

\item{prev}{Prevalance of the case in the general population.
Set it as 0 when it is not a case-control study.
Set it as a positive number when it is a case-control study.
If outcome.type="continuous": casecontrol=FALSE; prev=0
If outcome.type="binary":
          casecontrol=FALSE; prev=0 if it is not a case-control study
          casecontrol=TRUE; prev!=0 if it is a case-control study}

\item{printout}{Print the estimated coefficients for different paths. printout=FALSE}
}
\value{
list(CI_IE=CI_IE,CI_PM=CI_PM)
}
\description{
Calculate the 95\% confidence interval for indirect effect (IE) and percentage mediated using bootstrapping.
}
