% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CensoredMediatorCodes.R
\name{myIE}
\alias{myIE}
\title{Title myIE
## Description:        Calculate the indirect effect (IE) and percentage mediated. Called by "CensoredMediator".}
\usage{
myIE(alpha, beta, tmp, data, outcome.type)
}
\arguments{
\item{alpha}{Coefficients estimated for M~X.}

\item{beta}{Coefficients estimated for Y~X+M.}

\item{tmp}{AFT model residuals by the Kaplan-Meier estimator.}

\item{data}{Input dataset.}

\item{outcome.type}{The type of the outcome in the mediation model. outcome.type="continuous" or "binary".}
}
\value{
data.frame(IE,DE,TE,PM)
}
\description{
Title myIE
## Description:        Calculate the indirect effect (IE) and percentage mediated. Called by "CensoredMediator".
}
